Refactor selection card component and remove unused page
